weather_forecast/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==7.5.0
Django==4.2.11
djangorestframework==3.15.1
idna==3.7
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
